整合企业用户所属问题 AND 备份数据表-Lee

+++ application/admin/controller/Common.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * 权限认证
- * 
+ *
  * @author Lee<a605333742@gmail.com>
  * @date 2018-08-31
  */
@@ -15,7 +15,7 @@ class Common extends Controller{
     protected $auth;
     /*
      * 构造函数，继承父类构造函，数权限验证
-     * 
+     *
      * @return #
      */
     public function __construct(\think\App $app = null) {
@@ -25,13 +25,13 @@ class Common extends Controller{
         }
         if(session(config('USER_AUTH_KEY'))==NULL || empty(session(config('USER_AUTH_KEY')))){
             $this->redirect('/admin/Auth/index');
-        }       
+        }
         if(in_array(request()->module(),config('deny_module_list'))){
             return json(jsonData('无访问权限',300));
         }
         //权限验证
         $this->auth=new Auth();
-        if(strtolower(request()->controller())!='index'){            
+        if(strtolower(request()->controller())!='index'){
             $nodeList= $this->auth->ckeckAuth();
             $nodes=array();
             foreach ($nodeList as $vo){
@@ -42,7 +42,7 @@ class Common extends Controller{
             }
         }
     }
-    
+
     /*
      * 根据表单生成查询条件-进行列表过滤
      *
@@ -51,19 +51,19 @@ class Common extends Controller{
      * @return array
      */
     public function _search($model=''){
-        $map        =array ();
+        $map=array ();
         $fieldArray =$model->gettablefields()?$model->gettablefields():array();
         //进行精确查询的字段
-        $notLikeArray = array(
+        $notLikeArray =array(
             'p_id',
-        );
-        foreach ( $fieldArray as $key => $val ) {				
+       );
+        foreach($fieldArray as $key => $val ) {
             if (!input($val)|| input($val) == ''){
                 continue ;
             }
             if(strtotime(input($val))!==FALSE){
                 $map[]  =[$val,'between',timeTotimestamp(input($val))];
-            }elseif(in_array($val, array_diff($fieldArray,$notLikeArray))){
+            }elseif(in_array($val,array_diff($fieldArray,$notLikeArray))){
                 //特别指定一些字段进行模糊查询
                 $map[]  =[$val,"LIKE",'%'.trim(input($val)).'%'];
             } else {
@@ -71,9 +71,17 @@ class Common extends Controller{
                 $map[] =[$val,'=',input($val)];
             }
         }
+        if(session('user.is_company')!=NULL || session('user.is_company')!=0){
+            if(in_array("add_a_id",$fieldArray)){
+                $map[]=['add_a_id','=',[['=',session('user.id')],['=',session('user.add_a_id')],"OR"]];
+            }
+            if(in_array('user_id',$fieldArray)){
+                $map[]=['user_id','=',[['=',session('user.id')],['=',session('user.add_a_id')],"OR"]];
+            }
+        }
         return $map;
     }
-    
+
     /*
      * 根据表单生成查询条件-进行列表过滤
      *
@@ -86,7 +94,7 @@ class Common extends Controller{
      *
      * @return array    #
      */
-    public function _list($model, $map, $sortBy='id',$asc =TRUE, $countPk="id", $field="*") {        
+    public function _list($model,$map,$sortBy='id',$asc =TRUE,$countPk="id",$field="*") {
         //验证传参的有效性
         $asc = $asc?true:false;
         $countPk = $model->getPk()?$model->getPk():$countPk;
@@ -94,115 +102,114 @@ class Common extends Controller{
         $orderSign  =TRUE;
         $order  =input('_order');
         if(empty($order)){
-            $order  =$sortBy?$sortBy:$countPk;  
+            $order  =$sortBy?$sortBy:$countPk;
             $orderSign=$sortBy?(is_array($sortBy)?FALSE:TRUE):$orderSign;
         }
         //排序方式默认按照倒序排列
-        //接受 sort参数 0 表示倒序 非0都 表示正序        
+        //接受 sort参数 0 表示倒序 非0都 表示正序
         if($orderSign){
             $sort   =$asc? 'asc' : 'desc';
             $sort   =input('_sort')?(input('_sort') == 'asc' ? 'asc' : 'desc'):$sort;
-        }        
+        }
         //取得满足条件的记录数
         //过滤空字段
         $param=array_filter(array_merge(input(),$map));
-        $param['page']=1;        
+        $param['page']=1;
         if(isset($param['pageNum'])&&!empty($param['pageNum'])){
             $param['page']=$param['pageNum'];
             unset($param['pageNum']);
         }
         $listRows=input('numPerPage')?input('numPerPage'):20;
-        $voList = $model->where($map)->group($countPk)->field($field)->order($order.' '.$sort)->paginate($listRows,false,$param);         
+        $voList = $model->where($map)->group($countPk)->field($field)->order($order.' '.$sort)->paginate($listRows,false,$param);
         $sortImg =$sort; //排序图标
         $sortAlt =$sort == 'desc' ? '升序排列' : '倒序排列'; //排序提示
         $sort = $sort == 'desc' ? 1 : 0; //排序方式
         //模板赋值显示
-        $this->assign ( 'list', $voList );
-        $this->assign ( 'sort', $sort );
-        $this->assign ( 'order', $order );
-        $this->assign ( 'sortImg', $sortImg );
-        $this->assign ( 'sortType', $sortAlt );
-        $this->assign ( 'totalCount', $voList->total());
-        $this->assign ( 'numPerPage', $listRows );
-        $this->assign ( 'currentPage', !empty(input(config('VAR_PAGE')))?input(config('VAR_PAGE')):1);        
-    }  
-    
+        $this->assign('list',$voList);
+        $this->assign('sort',$sort);
+        $this->assign('order',$order);
+        $this->assign('sortImg',$sortImg);
+        $this->assign('sortType',$sortAlt);
+        $this->assign('totalCount',$voList->total());
+        $this->assign('numPerPage',$listRows);
+        $this->assign('currentPage',!empty(input(config('VAR_PAGE')))?input(config('VAR_PAGE')):1);
+    }
+
     /*
      * 显示首页也
-     * 
+     *
      * @return #
      */
     public function index($db='',$sort='id',$sortBy=TRUE){
-        $model  =$db?Db::name($db):Db::name(request()->controller());        
-        $map    =self::_search($model);        
+        $model  =$db?Db::name($db):Db::name(request()->controller());
+        $map    =self::_search($model);
         $map[]  =['status','=',1];
         $sort   =strtolower(request()->controller())=='node'?'path':$sort;
         self::_list($model,$map,$sort,$sortBy);
         return $this->fetch(request()->action());
     }
-    
+
     /*
      * 显示新增编辑页面
-     * 
+     *
      * @return #
      */
-    public function show(){        
-        if(input("sign")){
-            $this->assign("_sign",input("sign",'','code'));
-        }
+    public function show(){
+        if(input("sign")){$this->assign("_sign",input("sign",'','code'));}
         if(input("id")){
-            $model  =input('db')?Db::name(input('db','','code')):(Db::name(request()->controller())); 
+            $model  =input('db')?Db::name(input('db','','code')):(Db::name(request()->controller()));
             $id =is_numeric(input($model->getPk()))?input($model->getPk()):input($model->getPk(),'','code');
-            $this->assign('vo', $model->find($id));
+            $this->assign('vo',$model->find($id));
         }
         return $this->fetch();
-    }    
-    
+    }
+
     /*
      * 编辑数据
-     * 
-     * @return # 
+     *
+     * @return #
      */
     public function edit(){
         if(input('id')){
             return self::update()?json(jsonData('修改成功',201)):json(jsonData('修改失败',301));
         }
         return self::insert()?json(jsonData('新增成功',201)):json(jsonData('新增失败',301));
-    }    
-    
+    }
+
     /*
      * 新增数据
-     * 
+     *
      * @return #
      */
     public function insert(){
         $dbName =input("db")? input("db"): request()->controller();
-        $model  =Db::name($dbName);        
-        $field=$model->getTableFields(); 
+        $model  =Db::name($dbName);
         //字段过滤
         $insData=array();
         foreach (array_filter(input()) as $k=>$v){
-            if(in_array($k,$field)){
-                $insData[$k]=$v;
-            }
+            if(in_array($k,$model->getTableFields())){$insData[$k]=$v;}
+            //更新添加时间
+            if(in_array('add_time',$model->getTableFields())){$insData['add_time']  =time();}
+            //更新添加人
+            if(in_array('add_a_id',$model->getTableFields())){$insData['add_a_id']  =session('user.id');}
+             //更新修改时间
+            if(in_array('edit_time',$model->getTableFields())){$insData['edit_time']=time();}
+            //更新修改人
+            if(in_array('edit_a_id',$model->getTableFields())){$insData['edit_a_id']=session('user.id');}
         }
-        if (!$newId=$model->insertGetId($insData)) {            
-            return FALSE;
-        }
+        if (!$newId=$model->insertGetId($insData)){return FALSE;}
         $content= session('user.name').'在数据库 '.$dbName.' 新增了一条数据';
-        if($this->auth->addLog(1,$dbName,$content)){
-            return $newId;
-        }
+        if($this->auth->addLog(1,$dbName,$content)){return $newId;}
         return FALSE;
     }
-    
+
     /*
      * 更新数据
-     * 
+     *
      * $return #
      */
     public function update(){
-        $dbName =input("db")?input("db"):request()->controller();  
+        $dbName =input("db")?input("db"):request()->controller();
         $model  =Db::name($dbName);
         $model->startTrans();
         //判断是否有图片存在，标签图片是否更改
@@ -211,56 +218,52 @@ class Common extends Controller{
         if(self::invaImages($dbName,$condition)==FALSE){
             $model->rollback();
             return FALSE;
-        }       
+        }
         // 更新数据
         $upData=array();
         foreach(array_filter(input()) as $k=>$v){
-            if(in_array($k,$model->getTableFields())){
-                $upData[$k]=$v;
-            }
+            if(in_array($k,$model->getTableFields())){$upData[$k]=$v;}
+            //更新修改时间
+            if(in_array('edit_time',$model->getTableFields())){$upData['edit_time'] =time();}
+            //更新修改人
+            if(in_array('edit_a_id',$model->getTableFields())){$upData['edit_a_id'] =session('user.id');}
         }
         $result =$model->where($condition)->update($upData);
-        if (false == $result) {       
+        if (false == $result) {
             $model->rollback();
             return FALSE;
         }
-        $contents=session('user.name').'编辑的数据表'.$dbName.'中主键为'.$model->getPk().' 的数据';            
+        $contents=session('user.name').'编辑的数据表'.$dbName.'中主键为'.$model->getPk().' 的数据';
         if(!$this->auth->addLog(1,$dbName,$contents)){
             $model->rollback();
-            return FALSE;  
+            return FALSE;
         }
         $imgPath     ='/static/public/upload/';
-        if(session("oldImg")!=null && file_exists($imgPath.session("oldImg"))){
-            unlink($imgPath.session("oldImg"));
-        }            
-        if(session('oldSmallImg')!=null && file_exists($imgPath.session('oldSmallImg'))){
-            unlink($imgPath.session("oldSmallImg"));
-        }
+        if(session("oldImg")!=null && file_exists($imgPath.session("oldImg"))){unlink($imgPath.session("oldImg"));}
+        if(session('oldSmallImg')!=null && file_exists($imgPath.session('oldSmallImg'))){unlink($imgPath.session("oldSmallImg"));}
         $model->commit();
         return TRUE;
     }
-    
+
     /*
      * 删除数据
-     * 
+     *
      * @return #
      */
     public function del(){
         if(method_exists($this,'beforeDel')){
-            if($this->beforeDel()!==TRUE){
-                return json(jsonData($this->beforeDel(),300));
-            }
+            if($this->beforeDel()!==TRUE){return json(jsonData($this->beforeDel(),300));}
         }
         //删除指定记录
         $dbName =input("db")?input("db",'','code'):request()->controller();
-        $model  =Db::name($dbName);        
+        $model  =Db::name($dbName);
         $model->startTrans();
-        $pk = $model->getPk();
+        $pk =$model->getPk();
         $id =is_array(input($pk))?input($pk):input($pk,'','code');
         if (!$id) {
             $model->rollback();
             return json(jsonData('非法操作',300));
-        }       
+        }
         $condition  =[$pk=>$id];
         $setField   =$model->where($condition)->update(['status'=>0]);
         if($setField==FALSE){
@@ -271,7 +274,7 @@ class Common extends Controller{
         if($this->invaImages($dbName,$condition)==FALSE){
             $model->rollback();
             return json(jsonData("参数错误，删除失败",300));
-        } 
+        }
         $ids=is_array($id)?implode(',',$id):$id;
         $contents=session('user.name')."删除了数据表 ".$dbName. " 中主键为 ".$ids ." 的数据";
         if($this->auth->addLog(1,$dbName,$contents)){
@@ -280,12 +283,12 @@ class Common extends Controller{
         }
         $model->rollback();
         return json(jsonData("删除失败！"));
-                
+
     }
-    
+
     /*
      * 彻底删除记录（一般情况下不要使用）
-     * 
+     *
      * @return bool
      */
     public function delete($imagePath=''){
@@ -294,13 +297,9 @@ class Common extends Controller{
         $model  =Db::name($dbName);
         $pk     =$model->getPk();
         $id =is_array(input($pk))?input($pk):input($pk,'','code');
-        if(!$id){
-            return json(jsonData('非法操作',300));
-        }
+        if(!$id){  return json(jsonData('非法操作',300)); }
         $condition  =[$pk=>$id];
-        if (false == $model->where($condition)->delete()) {
-            return json(jsonData('删除失败!',300));
-        }
+        if (false == $model->where($condition)->delete()) { return json(jsonData('删除失败!',300)); }
         if($dbName=="Invaimg"){
             //如果操作的是清理图片表，那么就要执行清理
             $imgList=$model->where($condition)->column('imagepath');
@@ -317,10 +316,10 @@ class Common extends Controller{
         }
         return json(jsonData('删除成功!'));
     }
-    
+
     /*
      * 图片上传
-     * 
+     *
      * @return #
      */
     public function upImg($rootPath=''){
@@ -352,9 +351,7 @@ class Common extends Controller{
         }
         $rootPath=$rootPath?$rootPath:'/static/admin/file/';
         $info=$file->move($rootPath.$savePath);
-        if(!$info){
-           return FALSE;
-        }
+        if(!$info){return FALSE;}
         if(in_array($info->getExtension(),array("jpg,jpeg,png,gif"))){
             $image = \think\Image::open($rootPath.$savePath.$info->getSaveName());
             // 按照原图的比例生成一个最大为150*150的缩略图并保存为thumb.png
@@ -362,32 +359,28 @@ class Common extends Controller{
         }
         return $info?$info:FALSE;
     }
-    
+
     /*
      * 更改状态
-     * 
+     *
      * @return #
      */
-    public function changeStatus(){ 
+    public function changeStatus(){
         $dbName =input("db")?input("db","","code"):request()->controller();
         $model  =Db::name($dbName);
-        $pk = $model->getPk();
+        $pk =$model->getPk();
         $id =is_array(input($pk))?input($pk):input($pk,'','code');
-        if(!$id){
-            return json(jsonData('非法操作!',300));
-        }
+        if(!$id){ return json(jsonData('非法操作!',300)); }
         $condition  =[$pk=>$id];
         $result     =$model->where($condition)->setField('status',input("status","","code"));
-        $contents=session("user.name")."修改了数据表".$dbName.'中主键为'.$pk.'的数据状态';
-        if ($result !== false && $this->auth->addLog(1,$dbName,$contents)) {
-            return json(jsonData('状态修改成功!'));
-        }
+        $contents   =session("user.name")."修改了数据表".$dbName.'中主键为'.$pk.'的数据状态';
+        if ($result !== false && $this->auth->addLog(1,$dbName,$contents)) { return json(jsonData('状态修改成功!')); }
         return json(jsonData('状态修改失败!',300));
     }
-    
+
     /*
      * 修改和删除数据时，如果有图片，则把图片存在过期图片表中
-     * 
+     *
      * @return #
      */
     public function invaImages($dbName,$condition){
@@ -395,35 +388,17 @@ class Common extends Controller{
         $smallImg   =array();
         //查询被删除的数据中省份存在图片，如果存在图片，则把图片路径提出来，以便将来清理
         $delList=Db::name($dbName)->where($condition)->select();
-        if(!$delList){
-            return TRUE;
-        }
+        if(!$delList){  return TRUE;  }
         foreach($delList as $vo){
-            if(isset($vo['image']) && !empty($vo['image'])){
-                $imaPath[]  =$vo['image'];
-            }
-            if(isset ($vo['img']) && !empty ($vo['img'])){
-                $imaPath[]  =$vo['img'];
-            }
-            if(isset ($vo['icon'])&& !empty ($vo['icon'])){
-                $imaPath[]  =$vo['icon'];
-            }
-            if(isset ($vo['photo'])&& !empty ($vo['photo'])){
-                $imaPath[]  =$vo['photo'];
-            }
-            if(isset ($vo['logo'])&&!empty ($vo['logo'])){
-                $imaPath[]  =$vo['logo'];
-            }
-            if(isset($vo['small_image'])&& !empty($vo['small_image'])){
-                $smallImg[] =$vo['small_image'];
-            }
-            if(isset($vo['small_img'])&&!empty($vo['small_img'])){
-                $smallImg[] =$vo['small_img'];
-            }
-            if(isset($vo['small_photo'])&&!empty($vo['small_photo'])){
-                $smallImg[] =$vo['small_photo'];
-            }
-        }                
+            if(isset($vo['image']) && !empty($vo['image'])){$imaPath[]  =$vo['image'];  }
+            if(isset ($vo['img']) && !empty ($vo['img'])){$imaPath[]    =$vo['img']; }
+            if(isset ($vo['icon'])&& !empty ($vo['icon'])){$imaPath[]   =$vo['icon']; }
+            if(isset ($vo['photo'])&& !empty ($vo['photo'])){$imaPath[] =$vo['photo']; }
+            if(isset ($vo['logo'])&&!empty ($vo['logo'])){$imaPath[]    =$vo['logo']; }
+            if(isset($vo['small_image'])&& !empty($vo['small_image'])){$smallImg[]  =$vo['small_image']; }
+            if(isset($vo['small_img'])&&!empty($vo['small_img'])){$smallImg[]       =$vo['small_img']; }
+            if(isset($vo['small_photo'])&&!empty($vo['small_photo'])){$smallImg[]   =$vo['small_photo'];}
+        }
         $imaPaths   =array_merge($imaPath,$smallImg);
         if($imaPaths){
             $invaImg=Db::name("Invaimg");
@@ -437,6 +412,6 @@ class Common extends Controller{
             return TRUE;
         }
         return TRUE;
-    }    
+    }
 }
